--- a/utils/mock_ws_server/go.mod
+++ b/utils/mock_ws_server/go.mod
@@ -1,0 +1,5 @@
+module mock_ws_server
+
+go 1.23.4
+
+require github.com/gorilla/websocket v1.5.3 // indirect