### Problems
We were seeing low performance both in latency and number of messages processed in typescript.
This PR introduces a large performance improvement.

**Summary of Improvements:**

- Throughput improved by up to 50x (from ~500 msg/s to ~25,000 msg/s under extreme load).
- Latency reduced from ~1.7ms to near zero (Avg Lat: 1.73ms → -0.44ms).
- Dropped messages drastically reduced (from 2,481 to 101 at medium load; from 2,481 to 64,600 at extreme load, but with 50x more messages processed).


### How to test
- Added golang mock ws server
- Run mock server: `go run mock_ws_server.go --rate 50000`
- Update urls of exchange
```js
   const customWsUrl = 'ws://localhost:8080/ws';
    exchange.urls['api']['ws']['spot'] = customWsUrl;
    exchange.urls['api']['ws']['margin'] = customWsUrl;
    exchange.urls['api']['ws']['future'] = customWsUrl;
```

### Low load 100 msg/s
#### Before
=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 499       | 0       | 99.70         | 1.73        | 0.92        | 2.62        | 1.63        | 1.63        | 2.26        | 2.59        | 5005.00      

#### After
=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 499       | 0       | 99.76         | -0.05       | -0.79       | 0.18        | -0.03       | -0.03       | 0.01        | 0.11        | 5002.00  

### Medium load 5000 msg/s
#### Before
=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 2496      | 2495    | 499.10        | 1.48        | 1.24        | 3.04        | 1.49        | 1.49        | 1.53        | 1.54        | 5001.00      

#### After
=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 24512     | 101     | 4898.48       | -0.43       | -0.92       | 6.94        | -0.44       | -0.44       | -0.04       | 0.01        | 5004.00            


### Exteme load 50K msg/s

#### Before


=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 2482      | 2481    | 496.20        | 1.75        | 0.94        | 5.50        | 1.89        | 1.89        | 2.01        | 2.77        | 5002.00   

#### After
=== Benchmark Results Summary ===

Test Name              | Msgs Recv | Dropped | Thrput(msg/s) | Avg Lat(ms) | Min Lat(ms) | Max Lat(ms) | Med Lat(ms) | P50 Lat(ms) | P90 Lat(ms) | P99 Lat(ms) | Actual Duration(ms)
---------------------- | --------- | ------- | ------------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | ----------- | -------------------
watchTicker (original) | 124205    | 64600   | 24821.14      | -0.44       | -0.98       | 11.72       | -0.44       | -0.44       | -0.04       | 0.05        | 5004.00